requirements.txt
langgraph>=0.2.0
langgraph-cli[inmem]
langchain>=0.3.0
langchain-google-genai>=2.0.0
langsmith>=0.2.0
python-dotenv>=1.0.0

# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
apscheduler>=3.10.0

# Data visualization
jupyter>=1.0.0
matplotlib>=3.8.0
pandas>=2.1.0
seaborn>=0.13.0
plotly>=5.18.0
